+++ libs/runtime/src/identity.rs
@@ -4,7 +4,7 @@ impl pallet_identity::Config for crate::Runtime {
     type FieldDeposit = crate::ConstU128<{ crate::constants::IDENTITY_BYTE_DEPOOSIT }>;
     type ForceOrigin = frame_system::EnsureRoot<Self::AccountId>;
     type MaxAdditionalFields = crate::MaxAdditionalFields;
-    type MaxRegistrars = crate::ConstU32<{ crate::constants::MAX_AUTHORITIES }>;
+    type MaxRegistrars = crate::ConstU32<{ crate::constants::MAX_AUTHORITIES as u32 }>;
     type MaxSubAccounts = crate::ConstU32<{ crate::constants::IDENTITY_MAX_SUB_ACCOUNTS }>;
     type RegistrarOrigin = frame_system::EnsureRoot<Self::AccountId>;
     type RuntimeEvent = crate::RuntimeEvent;
